Refactor(Lotto) : Lotto 클래스의 위치를 lotto 패키지에서 domain 패키지로 변경
